+++ regression/python/if-else/main.py
@@ -1,5 +1,5 @@
 fail : bool = True
 if (fail == False):
-    result: int = 1/0
+    result:float = 1/0
 else:
-    result: int = 1/1
+    result:float = 1/1
